+++ lib/secure.js
@@ -1,8 +1,5 @@
 const sodium = require('sodium-native')
-const securePrompt = require('secure-prompt')
 const c = require('compact-encoding')
-
-const MIN_PASSWORD_LENGTH = 8
 
 const labelledKey = {
   preencode (state, k) {
@@ -77,9 +74,7 @@ const encryptedKey = {
 
 module.exports = {
   generateKeys,
-  sign,
-  confirmPassword,
-  readPassword
+  sign
 }
 
 function generateKeys (pwd) {
@@ -138,24 +133,6 @@ function generateKeys (pwd) {
   }
 }
 
-async function confirmPassword (pwd) {
-  const check = await readPassword('Confirm password: ')
-
-  if (pwd.byteLength !== check.byteLength) return false
-
-  sodium.sodium_mprotect_readonly(pwd)
-  sodium.sodium_mprotect_readwrite(check)
-
-  const cmp = sodium.sodium_memcmp(pwd, check)
-
-  sodium.sodium_memzero(check)
-  sodium.sodium_free(check)
-
-  sodium.sodium_mprotect_noaccess(pwd)
-
-  return cmp
-}
-
 function sign (signables, keyBuffer, pwd) {
   const { params, salt, payload } = c.decode(encryptedKey, keyBuffer)
 
@@ -194,21 +171,6 @@ function sign (signables, keyBuffer, pwd) {
   sodium.sodium_memzero(payload)
 
   return signatures
-}
-
-// function to accept password from user
-async function readPassword (prompt = 'Keypair password: ') {
-  process.stdout.write(prompt)
-
-  const pwd = await securePrompt()
-
-  console.log()
-
-  if (pwd.byteLength < MIN_PASSWORD_LENGTH) {
-    throw new Error(`Password must be at least ${MIN_PASSWORD_LENGTH} characters`)
-  }
-
-  return pwd
 }
 
 function xor (a, b) {
